tweaks on use of 'long'
git-svn-id: https://svn.r-project.org/R/trunk@42253 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/gnuwin32/sys-win32.c
@@ -130,7 +130,7 @@ typedef struct _FILETIME {
  
 void R_getProcTime(double *data)
 {
-    long  elapsed;
+    DWORD elapsed;
     double kernel, user;
     OSVERSIONINFO verinfo;
     /* This is in msec, but to clock-tick accuracy,
